% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2ranges.R
\name{bam2ranges}
\alias{bam2ranges}
\title{Import BAM file into GRanges}
\usage{
bam2ranges(file, bamindex = file, min.mapq = 10, pairedEndReads = FALSE,
  chromosomes = NULL)
}
\arguments{
\item{file}{Bamfile with aligned reads.}

\item{bamindex}{Bam-index file with or without the .bai ending. If this file does not exist it will be created and a warning is issued.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}
}
\description{
Import aligned reads from a BAM file into a \code{\link{GRanges}} object.
}
\author{
David Porubsky
}

