% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportUCSC.R
\name{exportUCSC}
\alias{exportUCSC}
\title{Generates a UCSC fromated bed file}
\usage{
exportUCSC(index, outputDirectory, fragments = NULL, segments = NULL,
  recombs = NULL, col = "103,139,139 243,165,97")
}
\arguments{
\item{index}{is used to index the bedfile(s)}

\item{outputDirectory}{is location to write bedfile(s)}

\item{fragments}{A \code{\link{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{bam2GRanges}}}

\item{segments}{A data.frame object with localized regions of 0 and 1 or mixture 0,1}

\item{recombs}{A data.frame object with coordinates of every state transition}
}
\description{
Write a bedfile from yeastSCElocatoR files for upload into a UCSC Genome browser
}
\author{
Ashley Sanders, David Porubsky
}

