% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSegmentsGAM.R
\name{findSegmentsGAM}
\alias{findSegmentsGAM}
\title{This function localize seqments of continous integer values}
\usage{
findSegmentsGAM(data = NULL)
}
\arguments{
\item{data}{A list of integer vectors of zeros and ones}
}
\description{
This function localize seqments of continous integer values
}
\author{
David Porubsky
}

