% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSegmentsHMM.R
\name{findSegmentsHMM}
\alias{findSegmentsHMM}
\title{This function localize seqments of continous integer values}
\usage{
findSegmentsHMM(data = NULL)
}
\arguments{
\item{data}{A list of integer vectors of zeros and ones}
}
\description{
This function localize seqments of continous integer values
}
\author{
David Porubsky
}

