% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSegmentsCBC.R
\name{findSegmentsCBS}
\alias{findSegmentsCBS}
\title{This function localize seqments of continous integer values}
\usage{
findSegmentsCBS(data = NULL, minSeg = 0, smooth = 3)
}
\arguments{
\item{data}{A list of integer vectors of zeros and ones}

\item{minSeg}{A minimal length of continuous segments of a certain value}

\item{smooth}{A number of subsequent integers considered as errors}
}
\description{
This function localize seqments of continous integer values
}
\author{
David Porubsky
}

