% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localizeSCE.R
\name{localizeSCE}
\alias{localizeSCE}
\title{Wrapper function for the YeastSCElocatoR package}
\usage{
localizeSCE(bamfilepath, dataDirectory = "./SCE_analysis", min.mapq = 10,
  minSeg = 5, smooth = 1, method = "CBC")
}
\arguments{
\item{bamfilepath}{Folder with BAM files to analyze}

\item{dataDirectory}{Folder to output the results. If it does not exist it will be created.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{minSeg}{A minimal length of continuous segments of a certain value}

\item{smooth}{A number of subsequent integers considered as errors}

\item{method}{}
}
\description{
Wrapper function for the YeastSCElocatoR package
}
\author{
David Porubsky
}

