% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localizeSCE.R
\name{localizeSCE}
\alias{localizeSCE}
\title{Wrapper function for the YeastSCElocatoR package}
\usage{
localizeSCE(inputDirectory, outputDirectory = "./SCE_analysis",
  chromosomes = NULL, min.mapq = 10, minSeg = 0, minSize = 0,
  smooth = 1, method = "CBS", hotspots = FALSE, read.len = 50)
}
\arguments{
\item{inputDirectory}{Folder with BAM files to analyze}

\item{outputDirectory}{Folder to output the results. If it does not exist it will be created.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{minSeg}{A minimal length of continuous segments of a certain value}

\item{smooth}{A number of subsequent integers considered as errors}
}
\description{
Wrapper function for the YeastSCElocatoR package
}
\author{
David Porubsky
}

